+++ CarWars/CarWars/Engine/Components/WeaponComponents/RailGunComponent.cpp
@@ -44,10 +44,12 @@ void RailGunComponent::Shoot(glm::vec3 position) {
                 line->SetPoint1(Transform::FromPx(gunHit.block.position));
                 auto tween = Effects::Instance().CreateTween<float, easing::Linear::easeNone>(1.f, 0.f, 0.1);
                 tween->SetUpdateCallback([line, rgTurret](float& value) mutable {
+                    if (StateManager::GetState() != GameState_Playing) return;
                     line->SetColor(glm::vec4(1.f, 0.f, 0.f, value));
                     line->SetPoint0(rgTurret->transform.GetGlobalPosition());
                 });
                 tween->SetFinishedCallback([bullet](float& value) mutable {
+                    if (StateManager::GetState() != GameState_Playing) return;
                     EntityManager::DestroyEntity(bullet);
                 });
                 tween->Start();
